fix(sequencer): clean automaton arrow — always returns Int step index
- SequencerAutomaton::step() now consistently returns ParamValue::Int
  for both transition and same-step cases
- Add table field to ServoTarget for index-based automata
- Add table field to ServoDef serialization

+++ rill-patchbay/src/automaton/sequencer.rs
@@ -180,8 +180,7 @@ impl Automaton for SequencerAutomaton {
             *internal = (next, time, new_dir, rng_state);
             ParamValue::Int(next as i32)
         } else {
-            // Same step — no change
-            ParamValue::Float(current_step as f32)
+            ParamValue::Int(current_step as i32)
         }
     }
